Tidy up the code
This commit changes the library to use a `vite.Config` for
configuration. The list of parameters is too sensitive to future
changes.

Furthermore, we add the ability to pass a public directory that will
server static files. In Vite apps, these are typically found in the
"public" directory. Public files are only served in development mode.
If there is no "PublicFS" specified in the configuration passed to the
handler, we will peek into the "FS" if there is a directory called
"public" and use that by default.

Finally, we will also initialize "ViteURL" in development mode with
"http://localhost:5173" if nothing else is specified.